scripts/policy-deep-research-agent/requirements.txt
accelerate
bitsandbytes
datasets
fastapi
numpy
openenv-core
pydantic>=2.0.0
pydantic-settings
python-dotenv
requests
torch
transformers
trl
uvicorn
langchain>=0.2.0
langchain-core
langgraph
langchain-openai
agentlightning
filelock
langsmith
sse-starlette
